UI: Removed Handles
Removed Handles from the UI as it was more cumbersome than anything else.
In theory, Widgets can be used after free (but not really cause Go's GC), and handles were supposed to prevent that.
In practice, it was not useful and made the API inconsistent.
So this refactored all the useful parts to an interface named "Container"